Replaced SWF Object with SoundManager2 library. More JS can probably be removed (jquery, jquery.flash).
Added CSS for page_as_playlist style.

Player is loading but the files are not. Seems to be a routing issue If we get this fixed then we should be on a roll.

In order for this to work in dev mode you have to change the Globa Security Settings of your flash player to allow local file access. To do this go to this URL and add the local ruxtape directory to the Trusted Files.

http://www.macromedia.com/support/documentation/en/flashplayer/help/settings_manager04.html