Made some interface changes
Editing the playlist and editing a song are more separated and clear this enables the song list to be more condensed for easier drag/drop reordering of long lists.

Also provided contingencies for long filenames in both the manager and the player.

Limited drag/drop to vertical(y) axis to avoid any layout breakages from all of the crazy dragging fools.